Run docker-compose as host user
When creating the `.go` volume on the host machine it was being created
using the default user, i.e, root.

This is not advisable as it will force the host user to either change
the permissions of the `.go` folder, run `sudo make test-component` or
delete the `.go` for subsquent executions.

Also we should not allow the default root user to create volumes on our
host machines as it is a security risk.